requirements.txt
asgiref==3.7.2
astroid==3.1.0
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.3
dill==0.3.8
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==24.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.1.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.4
uritemplate==4.1.1
urllib3==2.2.1
